Kbuild: Zero-init uninitialized locals for security
It has to be done this way since android-4.9 doesn't have any implementation
that both checks for compiler version as well as testing a specific flag in
Kconfig.

For clang 8+, -ftrivial-auto-var-init=zero is used, but the future of this
flag is uncertain. Comments of it are kanged from upstream commit.

Otherwise, make use of -fsanitize=local-init, which requires external patch
from HardenedOS that may or may not work with non-LTO builds so a guard is
needed for this case.

This significantly increases resulting kernel size depending on how many
uninitialized locals are being built.

Link: https://github.com/AndroidHardeningArchive/platform_external_clang/commit/776a0955ef6686d23a82d2e6a3cbd4a6a882c31c
Link: https://git.kernel.org/linus/f0fe00d4972a8cd4b98cc2c29758615e4d51cdfe
Signed-off-by: Albert I <kras@raphielgang.org>